Extract components from Pilots panel, and add initial test data
Extracted PilotsList plus child components
Extracted PilotDetails
Added a test data entry to Pilots state, and passed to other components
