+++ src/features/pilots/Pilot.js
@@ -1,0 +1,8 @@
+import {Model} from "redux-orm";
+
+
+export default class Pilot extends Model {
+
+}
+
+Pilot.modelName = "Pilot";
